pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "multimodal-fewshot-learning"
version = "0.1.0"
description = "Multi-modal Few-shot Learning with CLIP and Advanced Meta-Learning"
authors = [{name = "AI Research Team", email = "research@example.com"}]
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "torchaudio>=2.0.0",
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "tqdm>=4.62.0",
    "pyyaml>=6.0",
    "omegaconf>=2.2.0",
    "tensorboard>=2.8.0",
    "transformers>=4.21.0",
    "sentencepiece>=0.1.96",
    "accelerate>=0.12.0",
    "open_clip_torch>=2.20.0",
    "opencv-python>=4.6.0",
    "scikit-image>=0.19.0",
    "Pillow>=9.0.0",
    "sacrebleu>=2.0.0",
    "rouge-score>=0.1.2",
    "bert-score>=0.3.11",
    "nltk>=3.7",
    "matplotlib>=3.5.0",
    "plotly>=5.10.0",
    "seaborn>=0.11.0",
    "streamlit>=1.12.0",
    "gradio>=3.0.0",
    "fastapi>=0.78.0",
    "uvicorn>=0.18.0",
    "black>=22.0.0",
    "ruff>=0.0.200",
    "pytest>=7.0.0",
    "pre-commit>=2.17.0",
]

[project.optional-dependencies]
dev = [
    "wandb>=0.12.0",
    "mlflow>=1.28.0",
    "faiss-cpu>=1.7.0",
]

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
target-version = "py310"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*" = ["B011"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
